Conflicts:
	openpeer/core/cpp/core_Factory.cpp
	openpeer/core/cpp/core_ISystemMessage.cpp
	projects/xcode/hfcore/hfcore.xcodeproj/project.pbxproj
	projects/xcode/hfcoreTest_ios/hfcoreTest_ios.xcodeproj/project.pbxproj
	projects/xcode/hfcore_ios/hfcore_ios.xcodeproj/project.pbxproj
	projects/xcode/hfcoretest/hfcoretest.xcodeproj/project.pbxproj

+++ openpeer/core/cpp/core_MediaEngine.cpp
@@ -2267,9 +2267,9 @@ namespace openpeer
           return mError;
         }
         
-        unsigned int bytesSent;
+        size_t bytesSent;
         unsigned int packetsSent;
-        unsigned int bytesReceived;
+        size_t bytesReceived;
         unsigned int packetsReceived;
         
         mError = mVideoRtpRtcp->GetRTPStatistics(mVideoChannel, bytesSent, packetsSent, bytesReceived, packetsReceived);
@@ -3560,7 +3560,7 @@ namespace openpeer
       #pragma mark
 
       //-----------------------------------------------------------------------
-      int MediaEngine::RedirectTransport::SendPacket(int channel, const void *data, int len)
+      int MediaEngine::RedirectTransport::SendPacket(int channel, const void *data, size_t len)
       {
         Transport *transport = NULL;
         {
@@ -3576,7 +3576,7 @@ namespace openpeer
       }
 
       //-----------------------------------------------------------------------
-      int MediaEngine::RedirectTransport::SendRTCPPacket(int channel, const void *data, int len)
+      int MediaEngine::RedirectTransport::SendRTCPPacket(int channel, const void *data, size_t len)
       {
         Transport *transport = NULL;
         {
